Refactor on tests module structure
