+++ tests/fake_elasticsearch/test_delete.py
@@ -2,7 +2,7 @@
 
 from elasticsearch.exceptions import NotFoundError
 
-from tests.elasticsearch import TestElasticmock, INDEX_NAME, DOC_TYPE, BODY
+from tests import TestElasticmock, INDEX_NAME, DOC_TYPE, BODY
 
 
 class TestDelete(TestElasticmock):
