+++ tests/fake_elasticsearch/test_ping.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from tests.elasticsearch import TestElasticmock
+from tests import TestElasticmock
 
 
 class TestPing(TestElasticmock):
