+++ tests/fake_indices/__init__.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
